refactor: move common utility functions to test_utils
