+++ autonomi/Cargo.toml
@@ -16,7 +16,7 @@ data = []
 vault = ["data"]
 files = ["data"]
 fs = ["tokio/fs", "files"]
-local-discovery = ["sn_networking/local-discovery"]
+local-discovery = ["sn_networking/local-discovery", "test_utils/local-discovery"]
 registers = []
 
 [dependencies]
@@ -34,7 +34,6 @@ rmp-serde = "1.1.1"
 self_encryption = "~0.29.0"
 serde = { version = "1.0.133", features = ["derive", "rc"] }
 sn_networking = { path = "../sn_networking", version = "0.18.3" }
-sn_peers_acquisition = { path = "../sn_peers_acquisition", version = "0.5.2" }
 sn_protocol = { version = "0.17.10", path = "../sn_protocol" }
 sn_registers = { path = "../sn_registers", version = "0.3.20" }
 sn_transfers = { path = "../sn_transfers", version = "0.19.2" }
