+++ autonomi/tests/file.rs
@@ -8,14 +8,11 @@
 
 #![cfg(all(feature = "files", feature = "fs"))]
 
-mod common;
-
 use autonomi::Client;
-use common::peers_from_env;
 use eyre::Result;
 use sn_logging::LogBuilder;
 use std::time::Duration;
-use test_utils::evm::get_funded_wallet;
+use test_utils::{evm::get_funded_wallet, peers_from_env};
 use tokio::time::sleep;
 
 #[tokio::test]
