+++ autonomi/tests/wallet.rs
@@ -6,8 +6,6 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-mod common;
-
 use const_hex::traits::FromHex;
 use evmlib::common::{Address, Amount};
 use evmlib::utils::evm_network_from_env;
