+++ autonomi/tests/wasm.rs
@@ -12,22 +12,19 @@ use std::time::Duration;
 
 use autonomi::Client;
 use sn_networking::target_arch::sleep;
+use test_utils::{evm::get_funded_wallet, gen_random_data, peers_from_env};
 use wasm_bindgen_test::*;
-
-mod common;
 
 wasm_bindgen_test_configure!(run_in_browser);
 
 #[wasm_bindgen_test]
 async fn put() -> Result<(), Box<dyn std::error::Error>> {
-    common::enable_logging_wasm("sn_networking,autonomi,wasm");
+    enable_logging_wasm("sn_networking,autonomi,wasm");
 
-    let client = Client::connect(&common::peers_from_run_or_compile_time_env()?)
-        .await
-        .unwrap();
-    let wallet = test_utils::evm::get_funded_wallet();
+    let client = Client::connect(&peers_from_env()?).await.unwrap();
+    let wallet = get_funded_wallet();
 
-    let data = common::gen_random_data(1024 * 1024 * 2); // 2MiB
+    let data = gen_random_data(1024 * 1024 * 2); // 2MiB
     let addr = client.put(data.clone(), &wallet).await.unwrap();
 
     sleep(Duration::from_secs(2)).await;
@@ -36,4 +33,19 @@ async fn put() -> Result<(), Box<dyn std::error::Error>> {
     assert_eq!(data, data_fetched, "data fetched should match data put");
 
     Ok(())
+}
+
+fn enable_logging_wasm(directive: impl AsRef<str>) {
+    use tracing_subscriber::prelude::*;
+
+    console_error_panic_hook::set_once();
+
+    let fmt_layer = tracing_subscriber::fmt::layer()
+        .with_ansi(false) // Only partially supported across browsers
+        .without_time() // std::time is not available in browsers
+        .with_writer(tracing_web::MakeWebConsoleWriter::new()); // write events to the console
+    tracing_subscriber::registry()
+        .with(fmt_layer)
+        .with(tracing_subscriber::EnvFilter::new(directive))
+        .init();
 }
