+++ test_utils/Cargo.toml
@@ -9,11 +9,16 @@ readme = "README.md"
 repository = "https://github.com/maidsafe/safe_network"
 version = "0.4.6"
 
+[features]
+local-discovery = ["sn_peers_acquisition/local-discovery"]
 
 [dependencies]
+bytes = { version = "1.0.1", features = ["serde"] }
 color-eyre = "~0.6.2"
 dirs-next = "~2.0.0"
 evmlib = { path = "../evmlib", version = "0.1" }
 libp2p = { version = "0.54.1", features = ["identify", "kad"] }
+rand = "0.8.5"
 serde = { version = "1.0.133", features = [ "derive"]}
 serde_json = "1.0"
+sn_peers_acquisition = { path = "../sn_peers_acquisition", version = "0.5.2" }
