+++ test_utils/src/lib.rs
@@ -8,3 +8,30 @@
 
 pub mod evm;
 pub mod testnet;
+
+use bytes::Bytes;
+use color_eyre::eyre::Result;
+use libp2p::Multiaddr;
+use rand::Rng;
+use sn_peers_acquisition::parse_peer_addr;
+
+/// Generate random data of the given length.
+pub fn gen_random_data(len: usize) -> Bytes {
+    let mut data = vec![0u8; len];
+    rand::thread_rng().fill(&mut data[..]);
+    Bytes::from(data)
+}
+
+/// Parse the `SAFE_PEERS` env var into a list of Multiaddrs.
+///
+/// An empty `Vec` will be returned if the env var is not set or if local discovery is enabled.
+pub fn peers_from_env() -> Result<Vec<Multiaddr>> {
+    let bootstrap_peers = if cfg!(feature = "local-discovery") {
+        Ok(vec![])
+    } else if let Ok(peers_str) = std::env::var("SAFE_PEERS") {
+        peers_str.split(',').map(parse_peer_addr).collect()
+    } else {
+        Ok(vec![])
+    }?;
+    Ok(bootstrap_peers)
+}
